modifiche a HumidityFilter e HumidityFilterImpl

+++ ProgettoEsame/src/main/java/com/progetto/ProgettoEsame/filterAndStats/HumidityFilter.java
@@ -5,6 +5,6 @@ import java.util.Vector;
 
 public interface HumidityFilter {
 
-    public abstract Vector<Integer> getHumidityData (String cityName);
+    public abstract Vector<Long> getHumidityData (String cityName);
 
 }
